Split plot.py into multiple files. Updated flake8 to ignore __init__ files that only import things.
