+++ pints/plot/__init__.py
@@ -1,0 +1,15 @@
+#
+# Quick diagnostic plots.
+#
+# This file is part of PINTS (https://github.com/pints-team/pints/) which is
+# released under the BSD 3-clause license. See accompanying LICENSE.md for
+# copyright notice and full license details.
+#
+from ._autocorrelation import autocorrelation
+from ._function import function
+from ._function_between_points import function_between_points
+from ._histogram import histogram
+from ._pairwise import pairwise
+from ._series import series
+from ._trace import trace
+
